Flexibly register devices in a list
Every virtual device now exposes a default device_context, which is re-
ferenced, initialized and handled in ctroller.c.  The context holds all
the information of a virtual device, making initializing and exiting it
easier and less error prone.

A device_context contains the file descriptor that is associated with a
virtual device,  and function pointers to the respectice  creation- and
write- functions.  This makes it possible to create  more than one of a
specific device, (in theory) making it easier to connect  more than one
console.